MRADEV-665 Fix upload error
Fix error caused by only letting jpg images get resized. Resizer now allows images of any type to be resized. However, the error could not be completely eliminated. Part of the reason we were getting this is because of a bug caused by multer - a library we are using to upload images. In the error that occurs, the upload of the image works fine but when we attempt to access the location of the image it errors out - returning 'TODO: implement' rather than an image location.